added encoder.py
Refactored read_data into OOP style to avoid repeating all the function code for golomb, delta and theoretical new encodings. Also added a few new things like the ability to take random (seeded) samples of the data to encode, just to make it quicker to run and code that should truncate a block to the length of the longest binary in it. Will be adding stats analysis to each block and then to the sample as a whole.