Cleaned up plot block stats, fixed ratio stuff
Added titles and colours to the plots and made the histogram stuff work when plotting block stats. Adjusted how the compression ratio is obtained but this may need to be adjusted again to account for delta of delta or delta golomb encoding. Also uploaded two plots made of the stats plots