Fixed Golomb encoding
Overwrote the update bit difference method in Golomb class to reflect the fact that half of the encoded block is unary and half is remainders. Take max bit length of both halves and then multiply that by length of half block, then sum both of thsoe with codeword length. Takes space saving from -60 back to 2 - 10.