Can now run every direction for every file
Refactored load.cpp to load whole data file and then switch between them based on m_direction of Encoder. This reduces the amount of times we have to load data files which would add extra time to the execution. Also added filtering of data values outside of the normal range that was causing an error on file c3 11. Encodes every file in every direction from block size 3 -> 100 in 30 minutes for Delta encoding.